tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/cssfontfacerule-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSFontFaceRule - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssfontfacerule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSFontFaceRule' in window);

      }, '\'CSSFontFaceRule\' in window');

      test(function(){

        assert_true(CSSFontFaceRule instanceof Function);

      }, 'CSSFontFaceRule instanceof Function');

      test(function(){

        assert_true('prototype' in CSSFontFaceRule);

      }, '\'prototype\' in CSSFontFaceRule');

      test(function(){

        assert_equals(CSSFontFaceRule.prototype.toString(), '[object CSSFontFaceRulePrototype]');

      }, 'CSSFontFaceRule.prototype.toString() == \'[object CSSFontFaceRulePrototype]\'');

      test(function(){

        assert_true('constructor' in CSSFontFaceRule.prototype);

      }, '\'constructor\' in CSSFontFaceRule.prototype');

      test(function(){

        assert_equals(CSSFontFaceRule.prototype.constructor,CSSFontFaceRule);

      }, 'CSSFontFaceRule.prototype.constructor === CSSFontFaceRule');

    ]]>
    </script>
  </body>
</html>
